Feature/hooks (#362)
* [追加]パッケージを追加

* [追加]権限のハンドラを追加

* [追加]JS実行機能を追加

* [追加]マニフェストファイルを追加

* [追加]リリースチェック機能を追加

* [変更]プロパティー名を変更

* [追加]アドオンのDB側処理を追加

* [追加]マニフェストの読み込み機能を追加

* [追加]アドオンのインストーラーを追加

* [変更]アップデートに対応

* [追加]アドオンの実行機構を追加

* [追加]アドオンウィンドウを追加

* [変更]DIを解決

* [追加]ページを追加

* [変更]コンテクストの責任を整理

* [追加]複窓を禁止

* [追加]インストールページを追加

* [追加]アドオンインストールクラスを追加

* [追加]読み込みページを追加

* [追加]アドオン情報読み込みページの処理を追加

* [変更]読み込みページを整理

* [追加]インストールページを追加

* [修正]アドオンの読み込みを修正

* [追加]開発者設定を追加

* [追加]コードの読み込みを追加

* [追加]出力APIを追加

* [変更]移動動作を変更

* [追加]アンインストール機能を追加

* [追加]アップデート機能を追加

* [追加]デバッグ機能を追加

* [追加]デバッグ設定を追加

* [追加]ページ解析アドオンを作成

* [追加]インストールのキャンセル機能を追加

* アドオン管理ウィンドウの一部をページ化

* [追加]ホスト権限を追加

* [修正]ホスト権限の読み取りを修正

* [追加]fetch APIを追加

* [追加]ビルドスクリプトを追加

* [追加]Hooks APIを追加

* [変更]型定義を改善

* [変更]型定義ファイルを分割

* [追加]型定義を追加

* [削除]プロジェクトを移動

* [追加]型を公開

* [修正]動的な処理に対応

* [追加]htmlパース関数を追加

* [削除]不要なテストを削除

* [追加]Log APIを追加

* [追加]アイコン表示機能を追加

* [修正]正常にアドオンをアンインストールできない問題を修正

* [追加]再起動のポップアップを追加

* [変更]権限の説明を変更

* [修正]アップデートできない問題を修正

* [変更]終了時の警告表示をマテリアルデザインにした

* [修正]アイコンを更新できない問題を修正

* [修正]サムネを取得できない問題を修正

+++ Niconicome/Models/Domain/Niconico/Net/Json/API/Comment/Request/Comment.cs
@@ -29,7 +29,7 @@ namespace Niconicome.Models.Domain.Niconico.Net.Json.API.Comment.Request
         public string Version { get; set; } = string.Empty;
 
         [JsonPropertyName("fork")]
-        public int Fork { get; set; }
+        public long Fork { get; set; }
 
         [JsonPropertyName("language")]
         public int Language { get; set; }
@@ -71,7 +71,7 @@ namespace Niconicome.Models.Domain.Niconico.Net.Json.API.Comment.Request
         public string Thread { get; set; } = string.Empty;
 
         [JsonPropertyName("fork")]
-        public int Fork { get; set; }
+        public long Fork { get; set; }
 
         [JsonPropertyName("language")]
         public int Language { get; set; }
